requirements.txt
streamlit>=1.30.0
anthropic>=0.73.0
python-dotenv>=1.0.0
yfinance>=0.2.40
plotly>=5.18.0
pandas>=2.0.0
